refactor: 위치 수정
