+++ src/main/kotlin/edu/creamcommerce/application/product/usecase/CreateProductUseCase.kt
@@ -3,7 +3,7 @@ package edu.creamcommerce.application.product.usecase
 import edu.creamcommerce.application.product.dto.ProductDto
 import edu.creamcommerce.application.product.dto.command.CreateProductCommand
 import edu.creamcommerce.application.product.dto.toDto
-import edu.creamcommerce.domain.product.Money
+import edu.creamcommerce.domain.common.Money
 import edu.creamcommerce.domain.product.Product
 import edu.creamcommerce.domain.product.ProductOption
 import edu.creamcommerce.domain.product.ProductRepository
