+++ src/main/kotlin/edu/creamcommerce/domain/product/Product.kt
@@ -1,6 +1,6 @@
 package edu.creamcommerce.domain.product
 
-import java.math.BigDecimal
+import edu.creamcommerce.domain.common.Money
 import java.time.LocalDateTime
 import java.util.*
 
@@ -97,11 +97,3 @@ value class ProductId(val value: String) {
     }
 }
 
-@JvmInline
-value class Money(val amount: BigDecimal) {
-    constructor(amount: Int) : this(BigDecimal.valueOf(amount.toLong()))
-    
-    operator fun plus(other: Money): Money = Money(amount + other.amount)
-    operator fun minus(other: Money): Money = Money(amount - other.amount)
-    operator fun times(quantity: Int): Money = Money(amount * BigDecimal(quantity))
-}
