--- src/main/kotlin/edu/creamcommerce/interfaces/request/ProcessPaymentResponse.kt
@@ -1,7 +1,0 @@
-package edu.creamcommerce.interfaces.request
-
-data class ProcessPaymentResponse(
-    val paymentId: String,
-    val success: Boolean,
-    val message: String
-)
