+++ src/main/kotlin/edu/creamcommerce/interfaces/web/ApiResponse.kt
@@ -1,4 +1,4 @@
-package edu.creamcommerce.interfaces.response
+package edu.creamcommerce.interfaces.web
 
 import org.springframework.http.ResponseEntity
