+++ src/main/kotlin/edu/creamcommerce/interfaces/web/payment/ProcessPaymentRequest.kt
@@ -1,4 +1,4 @@
-package edu.creamcommerce.interfaces.request
+package edu.creamcommerce.interfaces.web.payment
 
 data class ProcessPaymentRequest(
     val orderId: String
