+++ src/test/kotlin/edu/creamcommerce/application/product/usecase/GetProductByIdUseCaseTest.kt
@@ -1,7 +1,7 @@
 package edu.creamcommerce.application.product.usecase
 
 import edu.creamcommerce.application.product.dto.toDto
-import edu.creamcommerce.domain.product.Money
+import edu.creamcommerce.domain.common.Money
 import edu.creamcommerce.domain.product.Product
 import edu.creamcommerce.domain.product.ProductId
 import edu.creamcommerce.domain.product.ProductRepository
